Bug Fix Amount & Default Photos

+++ database/migrations/2020_08_29_145328_create_expenses_table.php
@@ -19,7 +19,7 @@ class CreateExpensesTable extends Migration
             $table->string('reason');
             $table->string('status')->default('pending');
             $table->text('description');
-            $table->float('amount');
+            $table->int('amount');
             $table->string('receipt')->nullable();
             $table->timestamps();
         });
